Docker: Separated rubygems from source-files
Copying Gemfile & Gemfile.lock together with the source code in
Dockerfile causes bundler to install all gems on each change of source
code (like config.ru) even if there was no change in the rubygems
dependencies.

Facilitating docker image layer cache, it is possible to cache the bundle
install step. If only the source code is changed, rubygems are not
installed again, but the exiting docker image layer cache is used.